requirements.txt
firecrawl-py
phidata
pydantic
openai
python-dotenv
fastapi
uvicorn
pydantic-settings
beautifulsoup4